Command line options and usage reports
This is this first step in making Minder both a daemon and a tool
for collecting and displaying application usage data.

When passing the --daemon argument, it will run in the background
and collect data. Otherwise, you are able to display already-
collected data in both JSON and human-friendly formats.

More data reporting options are planned.